Refactor Alumni Card code
