+++ app/estudiants/page.tsx
@@ -1,6 +1,6 @@
 import { db } from "@/lib/db";
 
-import AlumniCard from "@/components/AlumniCard";
+import AlumniCard from "@/components/alumni_card/AlumniCard";
 
 export default async function Home() {
   return (
